+++ src/daf/message/text_based.py
@@ -120,7 +120,7 @@ class TextMESSAGE(BaseMESSAGE):
     def __init__(self,
                  start_period: Union[timedelta, int, None],
                  end_period: Union[int, timedelta],
-                 data: Union[Iterable[Union[str, discord.Embed, FILE]], _FunctionBaseCLASS, str, discord.Embed, FILE],
+                 data: Union[Iterable[Union[str, discord.Embed, FILE]], str, discord.Embed, FILE, _FunctionBaseCLASS],
                  channels: Union[Iterable[Union[int, discord.TextChannel, discord.Thread]], AutoCHANNEL],
                  mode: Optional[Literal["send", "edit", "clear-send"]] = "send",
                  start_in: Optional[Union[timedelta, bool]] = timedelta(seconds=0),